fix(connect-wallet): improve credential lookup and passkey naming
- Add pagination with while loop to query all REGISTER_CREDENTIAL events until match found
- Filter client-side by exact credentialId and domain to handle backend array_contains behavior
- Add timestamp suffix to passkey display name for easier identification across devices
- Use random base64url challenge and explicit timeout in WebAuthn authentication
- temporary debug logs

+++ spirekey/src/resolvers/create-wallet.ts
@@ -53,7 +53,7 @@ export const createWallet = async (
     getRootkeyPasskeyName(networkId!),
   );
 
-  const tempPassword = crypto.getRandomValues(new Uint16Array(32));
+  const tempPassword = crypto.getRandomValues(new Uint8Array(32));
   const entropy = await crypto.subtle.digest('sha-256', Buffer.from(hex));
   const mnemonic = bip39.entropyToMnemonic(new Uint8Array(entropy), wordlist);
   const seed = await kadenaMnemonicToSeed(tempPassword, mnemonic);
